Implement signal processing application with GUI
- Added main application logic in app.py for loading, plotting, and manipulating discrete signals.
- Created Signal class in my_signal.py to represent and handle signal operations such as addition, subtraction, scaling, shifting, and folding.
- Integrated matplotlib for visualizing signals within a PySide6 GUI.
- Implemented functionality for saving results to files and validating signal operations through test cases.
- Updated requirements.txt with necessary dependencies for the application.
