+++ .gitignore
@@ -1,0 +1,2 @@
+.venv/
+Task 1 testcases and testing functions.rar
