+++ Signal1.txt
@@ -1,4 +1,3 @@
-
 12
 -4 -2
 -3 2
