+++ Signal2.txt
@@ -1,4 +1,3 @@
-
 10
 -3 1
 -2 0
