--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
