--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/Signal1.txt
@@ -1,14 +1,0 @@
-
-12
--4 -2
--3 2
--2 0
--1 2
-0 4
-1 6
-2 3
-3 1
-4 -1
-5 -3
-6 0
-7 2
