--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/Signal2.txt
@@ -1,12 +1,0 @@
-
-10
--3 1
--2 0
--1 -1
-0 3
-1 2
-2 1
-3 -3
-4 6
-5 8
-6 3
