--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/add.txt
@@ -1,14 +1,0 @@
-
-12
--4 -2
--3 3
--2 0
--1 1
-0 7
-1 8
-2 4
-3 -2
-4 5
-5 5
-6 3
-7 2
