--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/advance3.txt
@@ -1,14 +1,0 @@
-
-12
--7 -2
--6 2
--5 0
--4 2
--3 4
--2 6
--1 3
-0 1
-1 -1
-2 -3
-3 0
-4 2
