--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/delay3.txt
@@ -1,14 +1,0 @@
-
-12
--1 -2
-0 2
-1 0
-2 2
-3 4
-4 6
-5 3
-6 1
-7 -1
-8 -3
-9 0
-10 2
