--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/folding.txt
@@ -1,14 +1,0 @@
-
-12
--7 2
--6 0
--5 -3
--4 -1
--3 1
--2 3
--1 6
-0 4
-1 2
-2 0
-3 2
-4 -2
