--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/mul5.txt
@@ -1,14 +1,0 @@
-
-12
--4 -10
--3 10
--2 0
--1 10
-0 20
-1 30
-2 15
-3 5
-4 -5
-5 -15
-6 0
-7 10
