--- Task 1 testcases and testing functions/Task 1 testcases and testing functions/subtract.txt
@@ -1,14 +1,0 @@
-
-12
--4 -2
--3 1
--2 0
--1 3
-0 1
-1 4
-2 2
-3 4
-4 -7
-5 -11
-6 -3
-7 2
