+++ advance3.txt
@@ -1,14 +1,13 @@
-
 12
--7 -2
--6 2
--5 0
--4 2
--3 4
--2 6
--1 3
-0 1
-1 -1
-2 -3
-3 0
-4 2
+-7 -2.0
+-6 2.0
+-5 0.0
+-4 2.0
+-3 4.0
+-2 6.0
+-1 3.0
+0 1.0
+1 -1.0
+2 -3.0
+3 0.0
+4 2.0
