+++ delay3.txt
@@ -1,14 +1,13 @@
-
 12
--1 -2
-0 2
-1 0
-2 2
-3 4
-4 6
-5 3
-6 1
-7 -1
-8 -3
-9 0
-10 2
+-1 -2.0
+0 2.0
+1 0.0
+2 2.0
+3 4.0
+4 6.0
+5 3.0
+6 1.0
+7 -1.0
+8 -3.0
+9 0.0
+10 2.0
