+++ folding.txt
@@ -1,14 +1,13 @@
-
 12
--7 2
--6 0
--5 -3
--4 -1
--3 1
--2 3
--1 6
-0 4
-1 2
-2 0
-3 2
-4 -2
+-7 2.0
+-6 0.0
+-5 -3.0
+-4 -1.0
+-3 1.0
+-2 3.0
+-1 6.0
+0 4.0
+1 2.0
+2 0.0
+3 2.0
+4 -2.0
