+++ mul5.txt
@@ -1,14 +1,13 @@
-
 12
--4 -10
--3 10
--2 0
--1 10
-0 20
-1 30
-2 15
-3 5
-4 -5
-5 -15
-6 0
-7 10
+-4 -10.0
+-3 10.0
+-2 0.0
+-1 10.0
+0 20.0
+1 30.0
+2 15.0
+3 5.0
+4 -5.0
+5 -15.0
+6 0.0
+7 10.0
